externalize utility functions
